requirements.txt
aiogram==3.4.1
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
